requirements.txt
fastapi
uvicorn
gunicorn

sqlalchemy
python-dotenv
email-validator
pydantic[email]
bcrypt
passlib
python-jose
python-multipart